+++ sdk/cards/ios/MoEReactNativeCardsHandler.m
@@ -7,9 +7,8 @@
 
 #import <Foundation/Foundation.h>
 #import "MoEReactNativeCardsHandler.h"
-#import "MoEngageReactUtils.h"
 #import "MoEngageCardsReactUtil.h"
-#import "MoEngageReactConstants.h"
+#import "MoEngageCardsReactConstants.h"
 
 @import MoEngagePluginCards;
 
@@ -29,7 +28,7 @@
 
 #pragma mark- Initialize methods
 -(void)initialize:(NSString*)payload {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] initialize:jsonPayload];
     [[MoEngagePluginCardsBridge sharedInstance] setSyncListener:jsonPayload];
     [[MoEngagePluginCardsBridge sharedInstance] setSyncEventListnerDelegate:self];
@@ -37,49 +36,49 @@
 
 #pragma mark- Callback methods
 -(void)isAllCategoryEnabled:(NSString *) payload resolve:(RCTPromiseResolveBlock) resolve reject:(RCTPromiseRejectBlock)reject {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] isAllCategoryEnabled:jsonPayload completionHandler:^(NSDictionary<NSString *,id> * _Nonnull cardPayload) {
         [MoEngageCardsReactUtil handleDataToReact:cardPayload rejecter:reject resolver:resolve];
     }];
 }
 
 -(void)getCardsCategories:(NSString *) payload resolve:(RCTPromiseResolveBlock) resolve reject:(RCTPromiseRejectBlock)reject {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] getCardsCategories:jsonPayload completionHandler:^(NSDictionary<NSString *,id> * _Nonnull cardPayload) {
         [MoEngageCardsReactUtil handleDataToReact:cardPayload rejecter:reject resolver:resolve];
     }];
 }
 
 -(void)getCardsInfo:(NSString *) payload resolve:(RCTPromiseResolveBlock) resolve reject:(RCTPromiseRejectBlock)reject {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] getCardsInfo:jsonPayload completionHandler:^(NSDictionary<NSString *,id> * _Nonnull cardPayload) {
         [MoEngageCardsReactUtil handleDataToReact:cardPayload rejecter:reject resolver:resolve];
     }];
 }
 
 -(void)getCardsForCategory:(NSString *) payload resolve:(RCTPromiseResolveBlock) resolve reject:(RCTPromiseRejectBlock)reject {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] getCardsForCategory:jsonPayload completionHandler:^(NSDictionary<NSString *,id> * _Nonnull cardPayload) {
         [MoEngageCardsReactUtil handleDataToReact:cardPayload rejecter:reject resolver:resolve];
     }];
 }
 
 -(void)fetchCards:(NSString *) payload resolve:(RCTPromiseResolveBlock) resolve reject:(RCTPromiseRejectBlock)reject {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] fetchCards:jsonPayload completionHandler:^(NSDictionary<NSString *,id> * _Nonnull cardPayload) {
         [MoEngageCardsReactUtil handleDataToReact:cardPayload rejecter:reject resolver:resolve];
     }];
 }
 
 -(void)getNewCardsCount:(NSString *) payload resolve:(RCTPromiseResolveBlock) resolve reject:(RCTPromiseRejectBlock)reject {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] getNewCardsCount:jsonPayload completionHandler:^(NSDictionary<NSString *,id> * _Nonnull cardPayload) {
         [MoEngageCardsReactUtil handleDataToReact:cardPayload rejecter:reject resolver:resolve];
     }];
 }
 
 -(void)getUnClickedCardsCount:(NSString *) payload resolve:(RCTPromiseResolveBlock) resolve reject:(RCTPromiseRejectBlock)reject {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] getUnClickedCardsCount: jsonPayload completionHandler:^(NSDictionary<NSString *,id> * _Nonnull cardPayload) {
         [MoEngageCardsReactUtil handleDataToReact:cardPayload rejecter:reject resolver:resolve];
     }];
@@ -87,38 +86,38 @@
 
 #pragma mark- Listener methods
 -(void)refreshCards:(NSString *) payload {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] refreshCards:jsonPayload];
 }
 
 -(void)onCardSectionLoaded:(NSString *) payload {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] onCardsSectionLoaded:jsonPayload];
 }
 
 -(void)onCardSectionUnLoaded:(NSString *) payload {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] onCardsSectionUnLoaded:jsonPayload];
 }
 
 #pragma mark- Stats methods
 -(void)cardClicked:(NSString *) payload {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] cardClicked:jsonPayload];
 }
 
 -(void)cardDelivered:(NSString *) payload {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] cardDelivered:jsonPayload];
 }
 
 -(void)cardShown:(NSString *) payload {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] cardShown:jsonPayload];
 }
 
 -(void)deleteCards:(NSString *) payload {
-    NSDictionary* jsonPayload = [MoEngageReactUtils getJSONRepresentation:payload];
+    NSDictionary* jsonPayload = [MoEngageCardsReactUtil getJSONRepresentation:payload];
     [[MoEngagePluginCardsBridge sharedInstance] deleteCards:jsonPayload];
 }