+++ sdk/cards/ios/MoEngageCardsReactUtil.h
@@ -14,4 +14,5 @@
 @interface MoEngageCardsReactUtil : NSObject
 +(NSString* _Nullable)fetchSyncType:(enum MoEngageCardsSyncEventType)eventType;
 +(void)handleDataToReact:(NSDictionary<NSString *,id> * _Nonnull)cardPayload rejecter:(RCTPromiseRejectBlock _Nullable )rejecter resolver:(RCTPromiseResolveBlock _Nullable )resolver;
++(NSDictionary* _Nullable)getJSONRepresentation:(NSString* _Nullable)string;
 @end