+++ sdk/cards/ios/MoEngageCardsReactUtil.m
@@ -40,5 +40,10 @@
     }
 }
 
++(NSDictionary*)getJSONRepresentation:(NSString*)string {
+    NSData *data = [string dataUsingEncoding:NSUTF8StringEncoding];
+    NSDictionary *jsonOutput = [NSJSONSerialization JSONObjectWithData:data options:0 error:nil];
+    return jsonOutput;
+}
 
 @end