+++ sdk/inbox/ios/MoEReactNativeInboxUtil.h
@@ -1,0 +1,13 @@
+//
+//  MoEReactNativeInboxUtil.h
+//  ReactNativeMoEngageInbox
+//
+//  Created by Rakshitha . on 26/12/24.
+//
+
+
+@import MoEngagePluginInbox;
+
+@interface MoEReactNativeInboxUtil : NSObject
++(NSDictionary*)getJSONRepresentation:(NSString*)string;
+@end